for json unflattening, use BigInt

+++ lift-base/lift-json/src/main/scala/net/liftweb/json/Extraction.scala
@@ -123,7 +123,7 @@ object Extraction {
         case "false" => JBool(false)
         case x @ _   => 
           if (value.charAt(0).isDigit) {
-            if (value.indexOf('.') == -1) JInt(value.toInt) else JDouble(value.toDouble)
+            if (value.indexOf('.') == -1) JInt(BigInt(value)) else JDouble(value.toDouble)
           }
           else JString(JsonParser.unquote(value.substring(1)))
       }
